swup を外す Fixes #639
