+++ package.json
@@ -1,17 +1,10 @@
 {
   "dependencies": {
-    "@mashvp/swup-html-lang-plugin": "^1.1.0",
-    "@swup/body-class-plugin": "^1.0.2",
-    "@swup/head-plugin": "^1.2.0",
-    "@swup/scripts-plugin": "^1.1.0",
-    "@swup/scroll-plugin": "^1.1.0",
     "@tweenjs/tween.js": "^18.6.4",
     "csshake": "^1.7.0",
     "gsap": "^3.10.4",
     "lil-gui": "^0.17.0",
     "swiper": "^8.3.0",
-    "swup": "^2.0.14",
-    "swup-morph-plugin": "^0.2.0",
     "three": "^0.141.0"
   },
   "devDependencies": {
