+++ themes/andaze/layouts/index-baseof.html
@@ -20,7 +20,7 @@
     <div class="absolute z-10 w-6 h-6 bg-white rounded-full animate-ping inset-1/2" id="loading_icon"></div>
   </div>
 
-  <div id="swup" class="transition-fade">
+  <div class="transition-fade">
     {{- block "main" . }}{{- end }}
   </div>
   {{- partial "footer.html" . -}}
